Grafana: Set Data Source references in Dashboard to Data Source names
This is done because the current Grafana Dashboard JSON has Data Sources referenced by IDs.
Those IDs are always generated new when Data Sources are added to Grafana, requiring an update
of IDs in the JSON. However, if Data Sources are referenced by name in the JSON, then it is
only important that Data Sources are named correctly. ReadMe is also updated to reflect
this change.